requirements.txt
fastapi
uvicorn
python-dotenv
injector
pydantic
